Remove tombi implementation
