+++ extension.toml
@@ -13,10 +13,6 @@ repository = "https://github.com/zed-extensions/toml"
 name = "Taplo"
 languages = ["TOML"]
 
-[language_servers.tombi]
-name = "Tombi"
-languages = ["TOML"]
-
 [grammars.toml]
 repository = "https://github.com/tree-sitter/tree-sitter-toml"
 commit = "342d9be207c2dba869b9967124c679b5e6fd0ebe"
