--- src/language_servers.rs
@@ -1,7 +1,0 @@
-mod config;
-mod taplo;
-mod tombi;
-mod util;
-
-pub use taplo::*;
-pub use tombi::*;
